requirements.txt
streamlit==1.37.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-option-menu==0.3.13
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
numpy==1.23.5
pandas>=1.3.0
scipy>=1.7.0
PyMySQL==1.1.1
matplotlib==3.9.1.post1
plotly==5.23.0
tqdm==4.67.1
db_dtypes==1.4.3
plotly
google-cloud-bigquery-storage==2.32.0